Global improvements

+++ install.sh
@@ -7,8 +7,6 @@ DOCKER_COMPOSE_VERSION=1.28.6
 NODEJS_VERSION=14
 PHP_VERSION=7.4
 KUBERNETES_VERSION=1.17.0
-GIT_EMAIL=${GIT_EMAIL:-"peru.remy@gmail.com"}
-GIT_NAME=${GIT_NAME:-"Rémy Peru"}
 
 # Create temp directory
 test -z "$TMPDIR" && TMPDIR="$(mktemp -d)"
@@ -31,36 +29,52 @@ mkdir -p "$HOME"/.fonts
 cp -r "$TMPDIR"/fonts "$HOME"/.fonts
 fc-cache -v
 
-# Install zsh
-export BINDIR=/usr/local/bin
+# Install zsh & plugin manager
 sudo apt-get install -y zsh
-curl -sL git.io/antibody | sudo -E sh -s
+curl -sfL git.io/antibody | sudo -E sh -s - -b /usr/local/bin
+
 cp -rf "$TMPDIR"/.zshrc "$HOME"/.zshrc
 
 # Change default shell to zsh
 sudo chsh $USER -s $(which zsh)
 
-# Install vim
+# vim
 sudo apt-get install -y vim
 
-# Install git
+# git
 sudo apt-get install -y git
-cp -rf "$TMPDIR"/.gitconfig "$HOME"/.gitconfig
-sed -E -i "s/(email =).*/\1 $GIT_EMAIL/" "$HOME"/.gitconfig
-sed -E -i "s/(name =).*/\1 $GIT_NAME/" "$HOME"/.gitconfig
 
-# Install hyper
-curl -sL https://github.com/zeit/hyper/releases/download/${HYPER_VERSION}/hyper_${HYPER_VERSION}_amd64.deb -o "$TMPDIR"/hyper.deb
-sudo apt-get install -y "$TMPDIR"/hyper.deb
+# Set git config if not already
+if test -z $(git config --global --get user.email); then
+  echo "What is your git email ?"
+  read git_email
+  echo "What is your git name ?"
+  read git_name
+
+  git config --global user.email $git_email
+  git config --global user.name $git_name
+fi
+git config --global core.editor vim
+
+# hyper
+echo "Installing hyper..."
+if test "$HYPER_VERSION" != $(hyper version) ; then
+  curl -L https://github.com/zeit/hyper/releases/download/${HYPER_VERSION}/hyper_${HYPER_VERSION}_amd64.deb -o "$TMPDIR"/hyper.deb
+  sudo apt-get install -y "$TMPDIR"/hyper.deb
+  echo "done"
+else
+  echo "already installed"
+fi
 cp -rf "$TMPDIR"/.hyper.js "$HOME"/.hyper.js
+echo "hyper configuration updated"
 
-# Install google-chrome
+# google-chrome
 wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
 echo "deb [arch=amd64] https://dl.google.com/linux/chrome/deb/ stable main" | sudo tee /etc/apt/sources.list.d/google-chrome.list
 sudo apt-get update
 sudo apt-get install -y google-chrome-stable
 
-# Install vscode
+# vscode
 if ! grep -q microsoft /proc/version; then # Not in wsl
   curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >"$TMPDIR"/microsoft.gpg
   sudo install -o root -g root -m 644 "$TMPDIR"/microsoft.gpg /etc/apt/trusted.gpg.d/
@@ -69,63 +83,95 @@ if ! grep -q microsoft /proc/version; then # Not in wsl
   sudo apt-get install -y code
 fi
 
-# Install gpg
+# gpg
 sudo apt-add-repository -y --update ppa:yubico/stable
 sudo apt-get install -y pcscd scdaemon gnupg2 pcsc-tools yubikey-manager
 mkdir -p "$HOME"/.gnupg
 cp -rf "$TMPDIR"/gpg-agent.conf "$HOME"/.gnupg/gpg-agent.conf
 
-# Install docker
+# docker
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
 echo "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list
 sudo apt-get update
 sudo apt-get install -y docker-ce docker-ce-cli containerd.io
 sudo usermod -aG docker "$USER"
 
-# Install docker-compose
+# docker-compose
+echo "Installing docker-compose..."
+if test "$DOCKER_COMPOSE_VERSION" != $(docker-compose version --short); then
 sudo curl -L https://github.com/docker/compose/releases/download/${DOCKER_COMPOSE_VERSION}/docker-compose-"$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose &&
   sudo chmod +x /usr/local/bin/docker-compose
+  echo "done"
+else
+  echo "already installed"
+fi
 
-# Install kubectl
-sudo curl -L https://storage.googleapis.com/kubernetes-release/release/v${KUBERNETES_VERSION}/bin/linux/amd64/kubectl -o /usr/local/bin/kubectl &&
+# kubectl
+echo "Installing kubectl..."
+if test "v${KUBERNETES_VERSION}" != $(kubectl version --client --short | awk '{print $3}'); then
+  sudo curl -L "https://dl.k8s.io/v${KUBERNETES_VERSION}/bin/linux/amd64/kubectl" -o /usr/local/bin/kubectl
   sudo chmod +x /usr/local/bin/kubectl
+  echo "done"
+else
+  echo "already installed"
+fi
 
-# Install kustomize
+# kustomize
+echo "Installing kustomize..."
 curl -s https://api.github.com/repos/kubernetes-sigs/kustomize/releases |
   grep browser_download |
-  grep linux |
+  grep linux_amd64 |
   cut -d '"' -f 4 |
   grep /kustomize/v |
-  sort | tail -n 1 |
+  sort |
+  tail -n 1 |
   xargs curl -L -o "$TMPDIR"/kustomize.tar.gz
 tar xzf "$TMPDIR"/kustomize.tar.gz -C "$TMPDIR"
 chmod +x "$TMPDIR"/kustomize
 sudo mv "$TMPDIR"/kustomize /usr/local/bin/kustomize
 
-# Install minikube
-curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64 &&
+# minikube
+echo "Installing minikube..."
+latest_version=$(curl -s https://api.github.com/repos/kubernetes/minikube/releases |
+ grep browser_download |
+ grep minikube-linux-amd64 |
+ cut -d '"' -f 4 |
+ head -n 1 |
+ awk -F'/' '{print $8}'
+)
+if test $latest_version != $(minikube version --short | awk '{print $3}'); then
+curl -LO "https://storage.googleapis.com/minikube/releases/${latest_version}/minikube-linux-amd64" &&
   sudo install minikube-linux-amd64 /usr/local/bin/minikube &&
   rm minikube-linux-amd64
+  echo "done"
+else
+  echo "already installed"
+fi
 
-# Install nodejs
+# nodejs
+echo "Installing nodejs..."
 curl -sL https://deb.nodesource.com/setup_${NODEJS_VERSION}.x | sudo -E bash -
 sudo apt-get install -y nodejs
 
-# Install yarn
+# yarn
+echo "Installing yarn..."
 curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
 echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
 sudo apt-get update
 sudo apt-get install -y yarn
 
-# Install php
+# php
+echo "Installing php..."
 sudo add-apt-repository -y --update ppa:ondrej/php
 sudo apt-get install -y php${PHP_VERSION} php${PHP_VERSION}-dev php${PHP_VERSION}-curl php${PHP_VERSION}-gd php${PHP_VERSION}-mbstring php${PHP_VERSION}-xml php${PHP_VERSION}-zip
 
-# Install composer
+# composer
+echo "Installing composer..."
 php -r "copy('https://getcomposer.org/installer', '$TMPDIR/composer-setup.php');"
 sudo php "$TMPDIR"/composer-setup.php --install-dir=/usr/local/bin --filename=composer
 
-# Install ansible
+# ansible
+echo "Installing ansible..."
 sudo apt-add-repository --yes --update ppa:ansible/ansible
 sudo apt install -y ansible
